pyproject.toml
[project]
name = "llm-finetuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.30",
    "datasets>=2.19",
    "evaluate>=0.4",
    "peft>=0.11",
    "pyyaml>=6.0",
    "sentencepiece>=0.2.1",
    "transformers>=4.42",
]
